Separate gen_macro.go
